OFZIPArchive: Restore accidentally dropped cast
FossilOrigin-Name: 6ae68302430bfbdd9b4c79f22d915a1d4eea630c2a026a9c6030048a14d03f29

+++ src/OFZIPArchive.m
@@ -894,7 +894,7 @@ seekOrThrowInvalidFormat(OFSeekableStream *stream,
 	} @catch (OFWriteFailedException *e) {
 		OFEnsure(e.bytesWritten <= length);
 
-		_bytesWritten += e.bytesWritten;
+		_bytesWritten += (unsigned long long)e.bytesWritten;
 		_CRC32 = OFCRC32(_CRC32, buffer, e.bytesWritten);
 
 		if (e.errNo == EWOULDBLOCK || e.errNo == EAGAIN)
@@ -903,7 +903,7 @@ seekOrThrowInvalidFormat(OFSeekableStream *stream,
 		@throw e;
 	}
 
-	_bytesWritten += length;
+	_bytesWritten += (unsigned long long)length;
 	_CRC32 = OFCRC32(_CRC32, buffer, length);
 
 	return length;
